requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
h11==0.13.0
idna==3.3
outcome==1.2.0
paho-mqtt==1.6.1
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
PyVirtualDisplay==3.0
requests==2.28.1
selenium==4.4.0
sniffio==1.2.0
sortedcontainers==2.4.0
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
webdriver-manager==3.8.3
wsproto==1.1.0
